Ensure node module paths can be built
If a developer requires a specific file from a node module we need to build
that bundle as if that file was the entry.  This should be looked at as a stop
gap and we really should coalese all the same files into a single stub so that
browserify can handle the deduping.