Builds given an array of trees
This concludes the work to make the pre-packager take an array of trees. This work encompassed a migration to th  rebuild API in broccoli, which allows us to simply pass the array trees in. While this default helps I had to intorduced a parsed tree cache which holds onto the inputPath for each one of the trees passed in.  Prior to this commit there was only one source directory due to a merge prior to entering.

This work also cleans up a lot of the information gathering and centralizes that logic as much as possible.