Remove "es:" requirement for ES6 code
Since Ember-CLI-Builder will discover all of the addons our app has we can safely cross reference the import's package against the parsed trees map. Addons and the App can be assumed that they were ES6, so if a package is not found in the map and does not contain a prefix we can safely assume that the import is just some ES6 module. We then pass the module through the es resolver.