Refactor to DRY up AllDependencies public API surface
This simply DRYs up the API surface. AllDependencies.update now knows how to deal with both new and existing dependencies. Prior to this commit we had a different method for realigning an existing graph. This also addresses my long standing TODO for the "last one wins" when updating a graph. This is fixed by introducing the concept of an idempotent update.