Use Babel for transpilation
Prior to this commit we used both esperanto and babel for compilation.  While this worked it left us in a place where we could not generate sourcemaps back into the source.  We also had 2 plugins that did module compilation.

There is a TODO here as I need to refactor the rest of the code account for the more detailed module info provided by Babel.