Recursively remove transitive dependencies
Prior to this commit I was not actually walking the graph and pruning away the transitive dependencies. When a file is removed we first compare that file against the entire graph, if no one has an edge into it we then look at it's imports and compare them against the graph. We recurse until there are either no more imports or we get to the point where an edge is running into a dependency.