Modify transitive test
To illustrate that the transitive path is being removed and only removed up to the point where there are no edges, I have modified the test to create a shared dependency between a stable node and a dropping node. This illustrates that nodes with edges do not get dropped when that are in the dropping transitive path.