Remove excessive enumeration within ES resolver
Prior to this commit we would blinding enter into the recursive graph resolution for ES2015 modules. While this does work it is extremely non-performant as you must execute a callback for each item in the array regardless if it had already been synced. This commit cleans up the conditions in which you enter resolution and also filters the imports to be only the ones that need to be synced.