Graph Refactor
This commit is an internal refactor to use a graph to represent the
imports inside of the application. The graph is now used as the source of
truth for what is actually synced to disk. The graph is also used to
determine what imports need to be pruned from the graph when the graph
mutates. Prior to this commit, we could not accurately prun away the the
the unreachable imports.

Reachability is determined by running the [postorder
algorithm](https://en.wikipedia.org/wiki/Tree_traversal#Post-order) for each of the root nodes. This effectively gives you all of the connected
nodes. You then simply get all of the nodes against the connected nodes.
What you are left with is the unreachable nodes. You then remove the
unreachable nodes from the graph.

Prior to this commit It was possible to have multiple versions of the
same library in the compiled output. This was due to the fact that we
would build a browserify bundle perch tree. This re-factor coaleases all
of the npm: imports into a single file and but can cache bust across
trees.

All APIs should be documented now.