Engine Boundary Algorithm
This introduces the algorithm to find the boundary of engines with in
the larger graph. The algorithm as follows:

1. Take each node in a root and look at it's outward edges
2. Determine if outward edge is point back at it's root or another root
3. Prune nodes that belong to another root
4. All local nodes found
5. Compare all local root nodes to find the intersection
6. Done

This ensures that things like tests and engines are resolved in the
graph but can safely be concated separately.

In a future iteration we should keep track of nodes we pruned away to
build a DAG of engines themselves to ensure a runtime all dependencies
are satisfied.